Version 1.0 functionality.  Integrating entire SHS Science(Core Physics) courses and activities into data model.
