+++ qpalx-elearning-domain/src/main/java/com/quaza/solutions/qpalx/elearning/domain/lms/adaptivelearning/LearningActivityE.java
@@ -8,6 +8,8 @@ public enum LearningActivityE {
 
     Quiz,
 
+    Video,
+
     Assesment,
 
     Assignment,
