+++ qpalx-elearning-web/src/main/resources/application.properties
@@ -17,9 +17,9 @@ mpower.live.api.url=https://app.mpowerpayments.com/api/v1/opr/create
 
 
 ########  Application Spring JPA Database Configuration  #######
-spring.datasource.url=jdbc:mysql://localhost/qpalxcom_staging
-spring.datasource.username=qpalxcom_webuser
-spring.datasource.password=HoLLer!#897gd
+spring.datasource.url=jdbc:mysql://localhost/QPalX
+spring.datasource.username=manyce400
+spring.datasource.password=bridge15
 spring.datasource.driver-class-name=com.mysql.jdbc.Driver
 spring.jpa.hibernate.naming_strategy=org.hibernate.cfg.EJB3NamingStrategy
 spring.jpa.hibernate.show_sql=true
