+++ qpalx-elearning-web/src/main/resources/log4j2.xml
@@ -5,19 +5,15 @@
             <PatternLayout pattern="%d{MMM-dd-yyyy HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n"/>
         </Console>
 
-        <File name="MAIN-APP-LOG-FILE" fileName="/home/qpalxcom/tomcat/logs/application.log">
-            <PatternLayout>
-                <Pattern>%d{MMM-dd-yyyy HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n</Pattern>
-            </PatternLayout>
-        </File>
+
     </Appenders>
     <Loggers>
         <Root level="INFO">
-            <AppenderRef ref="MAIN-APP-LOG-FILE"/>
+            <AppenderRef ref="Console"/>
         </Root>
 
         <Logger name="com.quaza.solutions" level="debug" additivity="false">
-            <AppenderRef ref="MAIN-APP-LOG-FILE" level="INFO" />
+            <AppenderRef ref="Console" level="INFO" />
         </Logger>
     </Loggers>
 </Configuration>
