+++ qpalx-elearning-web/src/main/resources/templates/qpalx-student/signup/proficiency.html
@@ -97,15 +97,15 @@
                     </select>
                 </div>
             </div>
-            <div class="col-md-3">
-                <div class="profi-box right-box"><img src="images/english.png" class="skill"/>
-                    <br/><br/>
-                    <p class="skill-text"><b>Vocational Studies</b></p>
-                    <select class="form-control profi-level" th:field="*{coreVocationalStudiesProficiencyLevel}">
-                        <option th:each="SimplifiedProficiencyRankE : ${SimplifiedProficiencyRanks}"  th:value="${SimplifiedProficiencyRankE}" th:text="${SimplifiedProficiencyRankE}"></option>
-                    </select>
-                </div>
-            </div>
+            <!--<div class="col-md-3">-->
+                <!--<div class="profi-box right-box"><img src="images/english.png" class="skill"/>-->
+                    <!--<br/><br/>-->
+                    <!--<p class="skill-text"><b>Vocational Studies</b></p>-->
+                    <!--<select class="form-control profi-level" th:field="*{coreVocationalStudiesProficiencyLevel}">-->
+                        <!--<option th:each="SimplifiedProficiencyRankE : ${SimplifiedProficiencyRanks}"  th:value="${SimplifiedProficiencyRankE}" th:text="${SimplifiedProficiencyRankE}"></option>-->
+                    <!--</select>-->
+                <!--</div>-->
+            <!--</div>-->
             <div class="col-md-1"></div>
         </div>
         <div class="clearfix bottom-space"></div>
